Created a script that starts the docker container for you using an env file
The name of the docker container is also the same name as the env file. you can now just do sudo Error: Failed to start Docker container 'example.com'.and a container will spunup easily from the env file. It seems more intuitive to do it this way, or im just lazy lol.